Don't require perspective to be a layout shift root.
If an element is a layout shift root, then it needs to be a paint
offset root also, or else the way we compute accumulated transform
offsets will fail when trying to cross the layout shift root.

Unfortunately the logic for the two look quite different, we should
unify in a subsequent patch.

Perspective was added as a layout shift root in a previous patch,
but there was I think not a particular reason to do so. If paint
offsets can accumulate across perspective, so can layout shift.

Fixed: 1157797

Change-Id: Ia6e123bbc9219ae03235af51a171a084ab8078c8