[sensors] Gravity sensor bindings and platform support
This CL implements blink bindings for Gravity sensor [1] and
adds Android, Win, Linux, CrOs platform support to get Gravity values.

Gravity Sensor can either be a physical sensor, in some phones for
example. In this case, the value is read directly from the sensor.
Otherwise if gravity sensor is not found, gravity value can be extracted
from the accelerometer.

The gravity fusion sensor implementation is similar to the linear
acceleration fusion sensor, since gravity extraction is a step in the
linear acceleration calculation.

[1] ED specification for Gravity Sensor:
https://w3c.github.io/accelerometer/#gravitysensor

https://chromestatus.com/feature/5384099747332096

Bug: 1163993
Change-Id: I756f2450176c2e3d96a9934bdadac133d57186e9